pyproject.toml
[project]
name = "Photo-Watermark"
version = "0.1.0"
description = "This project demonstrates how to perform various operations for adding watermarks to photos."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "piexif>=1.1.3",
    "pillow>=11.3.0",
    "pre-commit>=4.3.0",
    "typer>=0.19.2",
]


[tool.ruff]
target-version = "py312"
fix = true

[tool.ruff.lint]
extend-select = ["I"]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
